Minor GC & Occupancy improvements
- Change depths to shorts.  (We aren't going to see games with tree depth > 32K.)
- Always re-use score arrays.
- Remove unused TreeNode field.

+++ src/org/ggp/base/apps/validator/StateMachinePerformanceAnalyser.java
@@ -225,7 +225,7 @@ public class StateMachinePerformanceAnalyser
                              null);
 
           endTime = System.currentTimeMillis() + numSeconds*1000;
-          gameSearcher.startSearch(endTime, initialState, 0);
+          gameSearcher.startSearch(endTime, initialState, (short)0);
 
           if (ThreadControl.RUN_SYNCHRONOUSLY)
           {
